requirements/common.txt
Django==3.1.4
djangorestframework==3.12.2
drf-yasg==1.20.0
google-cloud-bigquery==2.5.0
pyarrow==2.0.0